navbar
added content to the navbar so that the user knows which game is being played and setup a logo so in future the user can click the logo to go to the homepage for "other" games.

burger icon added so that the user can use this to navigate through the pages and restructured html so that the sections apply correctly with the styling.