new game and betting functions
changes to the html and css for control bet and game areas made so that the buttons now fit on the same area and hide when not available to prevent issues with them being pressed whilst game running. the game now allows the player to add a bet before the game and lock this in to start the game to draw cards and after the result of the game a new one will start. The player also cannot bet more than his score which will be the remaining balance.