newgame break
fixed an issue when player draws so that the betting stage is bypassed, the pot remains to be won and a new set of cards is dealt to the table. organised code so it appears in a neater way and toggled the double button so it only appears when the player is able to double their bet. also added comments to help show what each function does for easier reference.