split game
added functions for the split game so that the option appears only when the player has two of the same value cards. so far the split card is stored elsewhere and the original hand is set to only the first card. bug with div being created for the second hand and bug with the double function not applying the winnings to the score.