bug fixes
fixed an issue with the dealer hitting when should stand in a split game by simplifying code. also amended issues outputted in terminal such as ul having a div rather than li and made changes to css js to target the li and other elements rather than div to get the same outcome as before.

colour changes to the hover in the nav sections to make more sense to the user at a glance.